lab4 finish

+++ calculator.h
@@ -3,9 +3,9 @@
 
 class Calculator
 {
-	public:
-		int Add (double, double);
-		int Sub (double, double);
+    public:
+    double Add (double, double);
+    double Sub (double, double);
 };
 
 #endif//CALCULATOR_H
